ref(): Created component namespace with Renderer class
