+++ lib/action_component/component.rb
@@ -1,0 +1,6 @@
+module ActionComponent
+  # Represents a component with a template, style and javascript file
+  class Component
+
+  end
+end
